Fixed bug when different bit_width are set for adc1 pins
ESP-IDF does not support different bit widths per channel on adc_unit_1. Since
the bit width is being set during configuration this would lead to the last
configured width to be used for all channels on adc unit 1, regardlessd of the
configured bit width configured for the channel during adc:start/1. This patch
sets the bit width again before taking a reading, since there is no way to be
sure that the bit width has not been changed since the initial configuration in
adc:start/1.

Signed-off-by: Winford <dwinford@pm.me>